ReportedArticleDTO 패키지 경로 변경 및 lombok 적용
reported_article.query -> reported_article.dto
