+++ src/main/java/org/thisdote/communityfordevelopers/reported_article/query/ReportedArticleMapper.java
@@ -1,6 +1,7 @@
 package org.thisdote.communityfordevelopers.reported_article.query;
 
 import org.apache.ibatis.annotations.Mapper;
+import org.thisdote.communityfordevelopers.reported_article.dto.ReportedArticleDTO;
 
 import java.util.List;
