+++ src/main/java/org/thisdote/communityfordevelopers/reported_article/query/ReportedArticleService.java
@@ -3,6 +3,7 @@ package org.thisdote.communityfordevelopers.reported_article.query;
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.thisdote.communityfordevelopers.reported_article.dto.ReportedArticleDTO;
 
 import java.util.List;
